23.04.21
dto 수정

+++ src/main/java/dto/Apply.java
@@ -8,7 +8,7 @@ public class Apply {
 	// 신청번호(시퀀스)
 	private long no;
 	// 참여인원
-	private long person;
+	private int person;
 	// 회원아이디(구매자)
 	private String memberid;
 	// 세션
